requirements.txt
google-cloud-bigquery==3.17.2
fastapi==0.111.0
uvicorn==0.30.1
pyyaml==6.0.1
fastapi
python-docx

